Move test/yomichan-vm.js to dev/vm.js
