+++ dev/dictionary-validate.js
@@ -18,7 +18,7 @@
 const fs = require('fs');
 const path = require('path');
 const {JSZip} = require('./util');
-const {VM} = require('../test/yomichan-vm');
+const {VM} = require('./vm');
 
 const vm = new VM();
 vm.execute([
